Update catTipoPermisosControlador.ts
Implementación de auditoria en los métodos del controlador
Implementación de transacciones

+++ servidor/src/controlador/catalogos/catTipoPermisosControlador.ts
@@ -1,8 +1,7 @@
 import { Request, Response } from 'express';
 import { QueryResult } from 'pg';
-import fs from 'fs';
+import AUDITORIA_CONTROLADOR from '../auditoria/auditoriaControlador';
 import pool from '../../database';
-const builder = require('xmlbuilder');
 
 class TipoPermisosControlador {
 
@@ -22,14 +21,61 @@ class TipoPermisosControlador {
   }
 
   // METODO PARA ELIMINAR REGISTROS
-  public async EliminarRegistros(req: Request, res: Response): Promise<void> {
-    const id = req.params.id;
-    await pool.query(
-      `
-      DELETE FROM cg_tipo_permisos WHERE id = $1
-      `
-      , [id]);
-    res.jsonp({ message: 'Registro eliminado.' });
+  public async EliminarRegistros(req: Request, res: Response): Promise<Response> {
+    try {
+      // TODO ANALIZAR COMOOBTENER USER_NAME E IP DESDE EL FRONT
+      const { user_name, ip } = req.body;
+      const id = req.params.id;
+
+      // INICIAR TRANSACCION
+      await pool.query('BEGIN');
+
+      // CONSULTAR DATOSORIGINALES
+      const rol = await pool.query('SELECT * FROM cg_tipo_permisos WHERE id = $1', [id]);
+      const [datosOriginales] = rol.rows;
+
+      if (!datosOriginales) {
+        // AUDITORIA
+        await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+          tabla: 'cg_tipo_permisos',
+          usuario: user_name,
+          accion: 'D',
+          datosOriginales: '',
+          datosNuevos: '',
+          ip,
+          observacion: `Error al eliminar el tipo de permiso con id ${id}`
+        });
+
+        // FINALIZAR TRANSACCION
+        await pool.query('COMMIT');
+        return res.status(404).jsonp({ message: 'Error al eliminar el registro.' });
+      }
+
+      await pool.query(
+        `
+        DELETE FROM cg_tipo_permisos WHERE id = $1
+        `
+        , [id]);
+
+      // AUDITORIA
+      await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+        tabla: 'cg_tipo_permisos',
+        usuario: user_name,
+        accion: 'D',
+        datosOriginales: JSON.stringify(datosOriginales),
+        datosNuevos: '',
+        ip,
+        observacion: null
+      });
+
+      // FINALIZAR TRANSACCION
+      await pool.query('COMMIT');
+      return res.jsonp({ message: 'Registro eliminado.' });
+    } catch (error) {
+      // FINALIZAR TRANSACCION
+      await pool.query('ROLLBACK');
+      return res.status(404).jsonp({ message: 'Error al eliminar el registro.' });
+    }
   }
 
   // METODO PARA LISTAR DATOS DE UN TIPO DE PERMISO
@@ -47,26 +93,78 @@ class TipoPermisosControlador {
   }
 
   // METODO PARA EDITAR REGISTRO
-  public async Editar(req: Request, res: Response): Promise<void> {
-    const id = req.params.id;
-    const { descripcion, tipo_descuento, num_dia_maximo, num_dia_anticipo, gene_justificacion, fec_validar, acce_empleado,
-      legalizar, almu_incluir, num_dia_justifica, num_hora_maximo, fecha_inicio, documento, contar_feriados, correo_crear,
-      correo_editar, correo_eliminar, correo_preautorizar, correo_autorizar, correo_negar, correo_legalizar, fecha_fin,
-      num_dia_anterior } = req.body;
-    await pool.query(
-      `
-      UPDATE cg_tipo_permisos SET descripcion = $1, tipo_descuento = $2, num_dia_maximo = $3, num_dia_anticipo = $4, 
-        gene_justificacion = $5, fec_validar = $6, acce_empleado = $7, legalizar = $8, almu_incluir = $9, 
-        num_dia_justifica = $10, num_hora_maximo = $11, fecha_inicio = $12, documento = $13, contar_feriados = $14, 
-        correo_crear = $15, correo_editar = $16, correo_eliminar = $17, correo_preautorizar = $18, correo_autorizar = $19, 
-        correo_negar = $20, correo_legalizar = $21, fecha_fin = $22, num_dia_anterior = $23
-      WHERE id = $24
-      `
-      , [descripcion, tipo_descuento, num_dia_maximo, num_dia_anticipo, gene_justificacion, fec_validar, acce_empleado,
+  public async Editar(req: Request, res: Response): Promise<Response> {
+    try {
+      const id = req.params.id;
+      const { descripcion, tipo_descuento, num_dia_maximo, num_dia_anticipo, gene_justificacion, fec_validar, acce_empleado,
         legalizar, almu_incluir, num_dia_justifica, num_hora_maximo, fecha_inicio, documento, contar_feriados, correo_crear,
         correo_editar, correo_eliminar, correo_preautorizar, correo_autorizar, correo_negar, correo_legalizar, fecha_fin,
-        num_dia_anterior, id]);
-    res.jsonp({ message: 'Registro actualizado.' });
+        num_dia_anterior, user_name, ip } = req.body;
+
+      // INICIAR TRANSACCION
+      await pool.query('BEGIN');
+
+      // CONSULTAR DATOSORIGINALES
+      const tipo = await pool.query('SELECT * FROM cg_tipo_permisos WHERE id = $1', [id]);
+      const [datosOriginales] = tipo.rows;
+
+      if (!datosOriginales) {
+        // AUDITORIA
+        await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+          tabla: 'cg_tipo_permisos',
+          usuario: user_name,
+          accion: 'U',
+          datosOriginales: '',
+          datosNuevos: '',
+          ip,
+          observacion: `Error al actualizar el tipo de permiso con id ${id}`
+        });
+
+        // FINALIZAR TRANSACCION
+        await pool.query('COMMIT');
+        return res.status(404).jsonp({ message: 'Error al actualizar el registro.' });
+      }
+
+      await pool.query(
+        `
+        UPDATE cg_tipo_permisos SET descripcion = $1, tipo_descuento = $2, num_dia_maximo = $3, num_dia_anticipo = $4, 
+          gene_justificacion = $5, fec_validar = $6, acce_empleado = $7, legalizar = $8, almu_incluir = $9, 
+          num_dia_justifica = $10, num_hora_maximo = $11, fecha_inicio = $12, documento = $13, contar_feriados = $14, 
+          correo_crear = $15, correo_editar = $16, correo_eliminar = $17, correo_preautorizar = $18, correo_autorizar = $19, 
+          correo_negar = $20, correo_legalizar = $21, fecha_fin = $22, num_dia_anterior = $23
+        WHERE id = $24
+        `
+        , [descripcion, tipo_descuento, num_dia_maximo, num_dia_anticipo, gene_justificacion, fec_validar, acce_empleado,
+          legalizar, almu_incluir, num_dia_justifica, num_hora_maximo, fecha_inicio, documento, contar_feriados, correo_crear,
+          correo_editar, correo_eliminar, correo_preautorizar, correo_autorizar, correo_negar, correo_legalizar, fecha_fin,
+          num_dia_anterior, id]);
+
+      // AUDITORIA
+      await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+        tabla: 'cg_tipo_permisos',
+        usuario: user_name,
+        accion: 'U',
+        datosOriginales: JSON.stringify(datosOriginales),
+        datosNuevos: `{descripcion: ${descripcion}, tipo_descuento: ${tipo_descuento}, num_dia_maximo: ${num_dia_maximo}, 
+                      num_dia_anticipo: ${num_dia_anticipo}, gene_justificacion: ${gene_justificacion}, fec_validar: ${fec_validar}, 
+                      acce_empleado: ${acce_empleado}, legalizar: ${legalizar}, almu_incluir: ${almu_incluir}, 
+                      num_dia_justifica: ${num_dia_justifica}, num_hora_maximo: ${num_hora_maximo}, fecha_inicio: ${fecha_inicio}, 
+                      documento: ${documento}, contar_feriados: ${contar_feriados}, correo_crear: ${correo_crear}, 
+                      correo_editar: ${correo_editar}, correo_eliminar: ${correo_eliminar}, correo_preautorizar: ${correo_preautorizar}, 
+                      correo_autorizar: ${correo_autorizar}, correo_negar: ${correo_negar}, correo_legalizar: ${correo_legalizar}, 
+                      fecha_fin: ${fecha_fin}, num_dia_anterior: ${num_dia_anterior}}`,
+        ip,
+        observacion: null
+      });
+
+      // FINALIZAR TRANSACCION
+      await pool.query('COMMIT');
+      return res.jsonp({ message: 'Registro actualizado.' });
+    } catch (error) {
+      // FINALIZAR TRANSACCION
+      await pool.query('ROLLBACK');
+      return res.status(404).jsonp({ message: 'Error al actualizar el registro.' });
+    }
   }
 
   // METODO PARA CREAR REGISTRO DE TIPO DE PERMISO
@@ -75,7 +173,10 @@ class TipoPermisosControlador {
       const { descripcion, tipo_descuento, num_dia_maximo, num_dia_anticipo, gene_justificacion, fec_validar, acce_empleado,
         legalizar, almu_incluir, num_dia_justifica, num_hora_maximo, fecha_inicio, documento, contar_feriados, correo_crear,
         correo_editar, correo_eliminar, correo_preautorizar, correo_autorizar, correo_negar, correo_legalizar,
-        fecha_fin, num_dia_anterior } = req.body;
+        fecha_fin, num_dia_anterior, user_name, ip } = req.body;
+
+      // INICIAR TRANSACCION
+      await pool.query('BEGIN');
 
       const response: QueryResult = await pool.query(
         `
@@ -93,6 +194,20 @@ class TipoPermisosControlador {
 
       const [tipo] = response.rows;
 
+      // AUDITORIA
+      await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+        tabla: 'cg_tipo_permisos',
+        usuario: user_name,
+        accion: 'I',
+        datosOriginales: '',
+        datosNuevos: JSON.stringify(tipo),
+        ip,
+        observacion: null
+      });
+
+      // FINALIZAR TRANSACCION
+      await pool.query('COMMIT');
+
       if (tipo) {
         return res.status(200).jsonp(tipo)
       }
@@ -101,7 +216,8 @@ class TipoPermisosControlador {
       }
     }
     catch (error) {
-      console.log(error)
+      // REVERTIR TRANSACCION
+      await pool.query('ROLLBACK');
       return res.jsonp({ message: 'error' });
     }
   }
